requirements.txt
fastapi
uvicorn
tortoise-orm
aerich
asyncmy
aiomysql
pytest
passlib
bcrypt
cryptography
passlib[bcrypt]
python-jose[cryptography]
python-multipart